% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin a continuous variable}
\usage{
bin_variable(vec, edges, prefix = "B")
}
\arguments{
\item{vec}{Numeric vector to bin}

\item{edges}{Numeric vector of bin edges}

\item{prefix}{Character prefix for bin labels (default: "B")}
}
\value{
Factor of binned values
}
\description{
Bin a continuous variable
}
