% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_default_bin_edges}
\alias{create_default_bin_edges}
\title{Create default bin edges for covariates}
\usage{
create_default_bin_edges()
}
\value{
List of bin edges for each covariate
}
\description{
Create default bin edges for covariates
}
