% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estiMINT-package.R
\docType{package}
\name{estiMINT-package}
\alias{estiMINT}
\alias{estiMINT-package}
\title{estiMINT: Build and Evaluate EIR Prediction Models}
\description{
Tools for building XGBoost and Random Forest models to estimate Entomological Inoculation Rate (EIR) from malariasimulation data.
}
\author{
\strong{Maintainer}: Cosmo Santoni \email{cosmo.santoni@imperial.ac.uk}

}
\keyword{internal}
