% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_pretrained_case_models}
\alias{load_pretrained_case_models}
\title{Load the pretrained case prediction models that ship with \emph{estiMINT}}
\usage{
load_pretrained_case_models()
}
\description{
Load the pretrained case prediction models that ship with \emph{estiMINT}
}
