% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_case_diagnostics}
\alias{plot_case_diagnostics}
\title{Generate diagnostic plots for case predictions}
\usage{
plot_case_diagnostics(predictions_df, output_dir)
}
\arguments{
\item{predictions_df}{Data frame from compare_case_predictions}

\item{output_dir}{Character path to output directory}
}
\description{
Generate diagnostic plots for case predictions
}
