% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_case_predictions_by_year}
\alias{plot_case_predictions_by_year}
\title{Create scatter plots for case predictions by year}
\usage{
plot_case_predictions_by_year(y_true, predictions_list, years, output_dir)
}
\arguments{
\item{y_true}{Numeric vector of true case values}

\item{predictions_list}{List of prediction vectors}

\item{years}{Numeric vector of years}

\item{output_dir}{Character path to output directory}
}
\description{
Create scatter plots for case predictions by year
}
