% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_predictions_by_prev_rank}
\alias{plot_predictions_by_prev_rank}
\title{Create scatter plots colored by prevalence rank}
\usage{
plot_predictions_by_prev_rank(y_true, predictions_list, prev_rank, output_dir)
}
\arguments{
\item{y_true}{Numeric vector of true values}

\item{predictions_list}{List of prediction vectors}

\item{prev_rank}{Numeric vector of prevalence ranks}

\item{output_dir}{Character path to output directory}
}
\description{
Create scatter plots colored by prevalence rank
}
