% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_predictions_combined}
\alias{plot_predictions_combined}
\title{Create combined prediction scatter plots}
\usage{
plot_predictions_combined(y_true, predictions_list, output_dir)
}
\arguments{
\item{y_true}{Numeric vector of true values}

\item{predictions_list}{List of prediction vectors}

\item{output_dir}{Character path to output directory}
}
\description{
Create combined prediction scatter plots
}
