% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_xgb_model}
\alias{run_xgb_model}
\title{Run xgb model with initial conditons}
\usage{
run_xgb_model(new_data, model = NULL)
}
\arguments{
\item{new_data}{Data frame with columns: prevalence (or prev_y9), dn0_use, Q0, phi_bednets,
seasonal, itn_use, irs_use}

\item{model}{An 'estiMINT_model' object; if NULL, tries global 'estiMINT_model'}
}
\value{
Numeric vector of calibrated EIR predictions
}
\description{
Run xgb model with initial conditons
}
