% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_cor}
\alias{safe_cor}
\title{Get user max threads for paralellisation}
\usage{
safe_cor(a, b)
}
\arguments{
\item{a}{standard deviation of covariate a}

\item{b}{standard deviation of covariate b}
}
\value{
Max number of user threads
}
\description{
Get user max threads for paralellisation
}
