% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{validate_data}
\alias{validate_data}
\title{Validate data dimensions and check for NAs}
\usage{
validate_data(X, y, data_name)
}
\arguments{
\item{X}{Feature matrix}

\item{y}{Target vector}

\item{data_name}{Character label for the dataset}
}
\description{
Validate data dimensions and check for NAs
}
