% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{load_xgb_model}
\alias{load_xgb_model}
\title{Load model onto memory for usage}
\usage{
load_xgb_model(path)
}
\arguments{
\item{path}{Path to .rds file or to a models directory containing estiMINT_model.rds}
}
\value{
An 'estiMINT_model' object
}
\description{
Load model onto memory for usage
}
