% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_interim.R
\name{predict_annual_cases}
\alias{predict_annual_cases}
\title{Predict annual clinical cases per 1000}
\usage{
predict_annual_cases(model_obj, new_data, feature_cols)
}
\arguments{
\item{model_obj}{Model object from train_xgboost or train_random_forest}

\item{new_data}{Data frame with new observations}

\item{feature_cols}{Character vector of feature column names}
}
\value{
numeric vector of predictions
}
\description{
Predict annual clinical cases per 1000
}
